Phase two of moving function search: Naive move, just lifts the search feature into the search bar and leaves things mostly as is.
