+++ assets/panel/debugger.properties
@@ -505,6 +505,8 @@ watchExpressionsSeparatorLabel2=\u0020→
 # in the functions search panel as a separator between function's inferred name
 # and its real name (if available).
 functionSearchSeparatorLabel=←
+functionSearch.search.placeholder=Search Functions...
+functionSearch.search.key=O
 
 # LOCALIZATION NOTE (resumptionOrderPanelTitle): This is the text that appears
 # as a description in the notification panel popup, when multiple debuggers are
