+++ src/components/Editor/SearchBar.css
@@ -126,3 +126,34 @@
   font-weight: bold;
 }
 
+.function-search .autocomplete li:first-child {
+  border-top: none;
+}
+
+.function-search .autocomplete li {
+  border: none;
+  color: var(--theme-comment);
+  border-top: 1px solid var(--theme-splitter-color);
+  border-bottom: 1px solid var(--theme-splitter-color);
+  background-color: var(--theme-tab-toolbar-background);
+  padding: 4px;
+  margin: 0;
+  display: flex;
+}
+
+.function-search .autocomplete li.selected {
+  color: var(--theme-body-background);
+  background-color: var(--theme-selection-background);
+}
+
+.function-search .autocomplete li .title {
+  font-weight: bold;
+  flex-grow: 2;
+  flex-basis: 95%;
+}
+
+.function-search .autocomplete li .subtitle {
+  font-weight: lighter;
+  flex-grow: 1;
+  flex-basis: 5%;
+}
